move HtmlGraphWriter to views package
